Docker Project/mqttInit/requirements.txt
requests
http.client
paho.mqtt